feat: Add Blue Archive Puzzle Solver GUI
This commit adds a graphical user interface (GUI) to the Blue Archive Puzzle Solver program. The GUI allows users to interact with the program by adjusting the number of rows and columns in the hexagonal grid. The grid is displayed using PyQt5's QGraphicsScene and QGraphicsView classes. Users can also click on individual hexagonal tiles to change their color.